+++ idea/tests/org/jetbrains/kotlin/idea/script/TestScriptDefinitionContributors.kt
@@ -20,14 +20,6 @@ import org.jetbrains.kotlin.idea.core.script.ScriptDefinitionContributor
 import org.jetbrains.kotlin.idea.core.script.loadDefinitionsFromTemplates
 import java.io.File
 import kotlin.script.dependencies.Environment
-import kotlin.script.dependencies.ScriptContents
-import kotlin.script.experimental.dependencies.AsyncDependenciesResolver
-import kotlin.script.experimental.dependencies.DependenciesResolver
-import kotlin.script.experimental.dependencies.ScriptDependencies
-import kotlin.script.experimental.dependencies.asSuccess
-import kotlin.script.experimental.location.ScriptExpectedLocation
-import kotlin.script.experimental.location.ScriptExpectedLocations
-import kotlin.script.templates.ScriptTemplateDefinition
 
 
 class CustomScriptTemplateProvider(
@@ -49,24 +41,7 @@ class FromTextTemplateProvider(
     override val id = "Test"
     override fun getDefinitions() = loadDefinitionsFromTemplates(
             templateClassNames = listOf("org.jetbrains.kotlin.idea.script.Template"),
-            templateClasspath = emptyList(),
+            templateClasspath = listOfNotNull(environment["template-classes"] as? File),
             environment = environment
     )
 }
-
-
-class FromTextDependenciesResolver : AsyncDependenciesResolver {
-    @Suppress("UNCHECKED_CAST")
-    suspend override fun resolveAsync(scriptContents: ScriptContents, environment: Environment): DependenciesResolver.ResolveResult {
-        return ScriptDependencies(
-                classpath = (environment["classpath"] as? List<File>).orEmpty(),
-                imports = (environment["imports"] as? List<String>).orEmpty(),
-                sources = (environment["sources"] as? List<File>).orEmpty()
-        ).asSuccess()
-    }
-}
-
-@ScriptExpectedLocations([ScriptExpectedLocation.Everywhere])
-@Suppress("unused")
-@ScriptTemplateDefinition(FromTextDependenciesResolver::class, scriptFilePattern = "script.kts")
-class Template
